Added DeleteConfirmDialog

+++ frontend/src/components/recipe-in-list/recipe-in-list.component.jsx
@@ -1,26 +1,45 @@
 import './recipe-in-list.style.scss'
-import {Link, useNavigate} from "react-router-dom";
+import {Link} from "react-router-dom";
 import axios from "axios";
+import DialogDeleteRecipe from "../delete-recipe/delete-recipe.component";
+import {useState} from "react";
 
 
 const RecipeInList = ({setIsUpToDate, recipe}) => {
     const {id, name, description, image} = recipe;
-    let navigate = useNavigate();
 
-    const deleteRecipe = async (id) => {
-        await axios.delete(`http://localhost:8000/recipes/${id}/`)
-            .then(response => {
-                console.log(response.data);
-                navigate('/');
-                setIsUpToDate(false);
-            })
+    const [dialog, setDialog] = useState({
+        isOpen: false,
+    });
+
+    const handleDialog = (isOpen) => {
+        setDialog({
+            isOpen,
+        });
+    };
+
+    const handleDelete = () => {
+        handleDialog(true);
+    };
+
+    const areUSureDelete = async (choiceYes) => {
+        if (choiceYes) {
+            await axios.delete(`http://localhost:8000/recipes/${id}/`)
+                .then(response => {
+                    console.log(response.data);
+                    setIsUpToDate(false);
+                })
+            handleDialog(false);
+        } else {
+            handleDialog(false);
+        }
     };
 
     return (
         <div key={id} className="recipe-container">
             <Link to={"/recipes/" + id}>
                 <div className="image-container">
-                    <img src={image}></img>
+                    <img src={image}/>
                 </div>
                 <div className="text-container">
                     <h2>{name}</h2>
@@ -31,8 +50,13 @@ const RecipeInList = ({setIsUpToDate, recipe}) => {
                 <Link to={`/recipes/${id}/update`}>
                     <button className="button-style">Edit</button>
                 </Link>
-                <button className="button-style" onClick={() => deleteRecipe(id)}>Delete</button>
+                <button className="button-style" onClick={() => handleDelete()}>Delete</button>
             </div>
+            {dialog.isOpen && (
+                <DialogDeleteRecipe
+                    onDialog={areUSureDelete}
+                />
+            )}
         </div>
     );
 }
